+++ core/translate/planner.rs
@@ -1,5 +1,5 @@
 use super::plan::{
-    Aggregate, BTreeTableReference, Direction, Plan, ResultSetColumn, SourceOperator,
+    Aggregate, BTreeTableReference, Direction, GroupBy, Plan, ResultSetColumn, SourceOperator,
 };
 use crate::{function::Func, schema::Schema, util::normalize_ident, Result};
 use sqlite3_parser::ast::{self, FromClause, JoinType, ResultColumn};
@@ -19,9 +19,9 @@ impl OperatorIdCounter {
     }
 }
 
-fn resolve_aggregates(expr: &ast::Expr, aggs: &mut Vec<Aggregate>) {
+fn resolve_aggregates(expr: &ast::Expr, aggs: &mut Vec<Aggregate>) -> bool {
     if aggs.iter().any(|a| a.original_expr == *expr) {
-        return;
+        return true;
     }
     match expr {
         ast::Expr::FunctionCall { name, args, .. } => {
@@ -31,17 +31,22 @@ fn resolve_aggregates(expr: &ast::Expr, aggs: &mut Vec<Aggregate>) {
                 0
             };
             match Func::resolve_function(normalize_ident(name.0.as_str()).as_str(), args_count) {
-                Ok(Func::Agg(f)) => aggs.push(Aggregate {
-                    func: f,
-                    args: args.clone().unwrap_or_default(),
-                    original_expr: expr.clone(),
-                }),
+                Ok(Func::Agg(f)) => {
+                    aggs.push(Aggregate {
+                        func: f,
+                        args: args.clone().unwrap_or_default(),
+                        original_expr: expr.clone(),
+                    });
+                    true
+                }
                 _ => {
+                    let mut contains_aggregates = false;
                     if let Some(args) = args {
                         for arg in args.iter() {
-                            resolve_aggregates(arg, aggs);
+                            contains_aggregates |= resolve_aggregates(arg, aggs);
                         }
                     }
+                    contains_aggregates
                 }
             }
         }
@@ -53,15 +58,20 @@ fn resolve_aggregates(expr: &ast::Expr, aggs: &mut Vec<Aggregate>) {
                     func: f,
                     args: vec![],
                     original_expr: expr.clone(),
-                })
+                });
+                true
+            } else {
+                false
             }
         }
         ast::Expr::Binary(lhs, _, rhs) => {
-            resolve_aggregates(lhs, aggs);
-            resolve_aggregates(rhs, aggs);
+            let mut contains_aggregates = false;
+            contains_aggregates |= resolve_aggregates(lhs, aggs);
+            contains_aggregates |= resolve_aggregates(rhs, aggs);
+            contains_aggregates
         }
         // TODO: handle other expressions that may contain aggregates
-        _ => {}
+        _ => false,
     }
 }
 
@@ -271,19 +281,7 @@ pub fn prepare_select_plan<'a>(schema: &Schema, select: ast::Select) -> Result<P
             for column in columns.clone() {
                 match column {
                     ast::ResultColumn::Star => {
-                        for table_reference in plan.referenced_tables.iter() {
-                            for (idx, col) in table_reference.table.columns.iter().enumerate() {
-                                plan.result_columns.push(ResultSetColumn {
-                                    expr: ast::Expr::Column {
-                                        database: None, // TODO: support different databases
-                                        table: table_reference.table_index,
-                                        column: idx,
-                                        is_rowid_alias: col.primary_key,
-                                    },
-                                    contains_aggregates: false,
-                                });
-                            }
-                        }
+                        plan.source.select_star(&mut plan.result_columns);
                     }
                     ast::ResultColumn::TableStar(name) => {
                         let name_normalized = normalize_ident(name.0.as_str());
@@ -340,10 +338,8 @@ pub fn prepare_select_plan<'a>(schema: &Schema, select: ast::Select) -> Result<P
                                         });
                                     }
                                     Ok(_) => {
-                                        let cur_agg_count = aggregate_expressions.len();
-                                        resolve_aggregates(&expr, &mut aggregate_expressions);
                                         let contains_aggregates =
-                                            cur_agg_count != aggregate_expressions.len();
+                                            resolve_aggregates(&expr, &mut aggregate_expressions);
                                         plan.result_columns.push(ResultSetColumn {
                                             expr: expr.clone(),
                                             contains_aggregates,
@@ -380,10 +376,8 @@ pub fn prepare_select_plan<'a>(schema: &Schema, select: ast::Select) -> Result<P
                                 }
                             }
                             expr => {
-                                let cur_agg_count = aggregate_expressions.len();
-                                resolve_aggregates(expr, &mut aggregate_expressions);
                                 let contains_aggregates =
-                                    cur_agg_count != aggregate_expressions.len();
+                                    resolve_aggregates(expr, &mut aggregate_expressions);
                                 plan.result_columns.push(ResultSetColumn {
                                     expr: expr.clone(),
                                     contains_aggregates,
@@ -393,18 +387,37 @@ pub fn prepare_select_plan<'a>(schema: &Schema, select: ast::Select) -> Result<P
                     }
                 }
             }
-            if let Some(group_by) = group_by.as_mut() {
+            if let Some(mut group_by) = group_by {
                 for expr in group_by.exprs.iter_mut() {
                     bind_column_references(expr, &plan.referenced_tables)?;
                 }
-                if aggregate_expressions.is_empty() {
-                    crate::bail_parse_error!(
-                        "GROUP BY clause without aggregate functions is not allowed"
-                    );
-                }
+
+                plan.group_by = Some(GroupBy {
+                    exprs: group_by.exprs,
+                    having: if let Some(having) = group_by.having {
+                        let mut predicates = vec![];
+                        break_predicate_at_and_boundaries(having, &mut predicates);
+                        for expr in predicates.iter_mut() {
+                            bind_column_references(expr, &plan.referenced_tables)?;
+                            let contains_aggregates =
+                                resolve_aggregates(expr, &mut aggregate_expressions);
+                            if !contains_aggregates {
+                                // TODO: sqlite allows HAVING clauses with non aggregate expressions like
+                                // HAVING id = 5. We should support this too eventually (I guess).
+                                // sqlite3-parser does not support HAVING without group by though, so we'll
+                                // need to either make a PR or add it to our vendored version.
+                                crate::bail_parse_error!(
+                                    "HAVING clause must contain an aggregate function"
+                                );
+                            }
+                        }
+                        Some(predicates)
+                    } else {
+                        None
+                    },
+                });
             }
 
-            plan.group_by = group_by.map(|g| g.exprs);
             plan.aggregates = if aggregate_expressions.is_empty() {
                 None
             } else {
@@ -513,13 +526,14 @@ fn parse_from(
 
     let mut table_index = 1;
     for join in from.joins.unwrap_or_default().into_iter() {
-        let (right, outer, predicates) =
+        let (right, outer, using, predicates) =
             parse_join(schema, join, operator_id_counter, &mut tables, table_index)?;
         operator = SourceOperator::Join {
             left: Box::new(operator),
             right: Box::new(right),
             predicates,
             outer,
+            using,
             id: operator_id_counter.get_next_id(),
         };
         table_index += 1;
@@ -534,7 +548,12 @@ fn parse_join(
     operator_id_counter: &mut OperatorIdCounter,
     tables: &mut Vec<BTreeTableReference>,
     table_index: usize,
-) -> Result<(SourceOperator, bool, Option<Vec<ast::Expr>>)> {
+) -> Result<(
+    SourceOperator,
+    bool,
+    Option<ast::DistinctNames>,
+    Option<Vec<ast::Expr>>,
+)> {
     let ast::JoinedSelectTable {
         operator,
         table,
@@ -563,18 +582,62 @@ fn parse_join(
 
     tables.push(table.clone());
 
-    let outer = match operator {
+    let (outer, natural) = match operator {
         ast::JoinOperator::TypedJoin(Some(join_type)) => {
-            if join_type == JoinType::LEFT | JoinType::OUTER {
-                true
-            } else {
-                join_type == JoinType::RIGHT | JoinType::OUTER
-            }
+            let is_outer = join_type.contains(JoinType::OUTER);
+            let is_natural = join_type.contains(JoinType::NATURAL);
+            (is_outer, is_natural)
         }
-        _ => false,
+        _ => (false, false),
     };
 
+    let mut using = None;
     let mut predicates = None;
+
+    if natural && constraint.is_some() {
+        crate::bail_parse_error!("NATURAL JOIN cannot be combined with ON or USING clause");
+    }
+
+    let constraint = if natural {
+        // NATURAL JOIN is first transformed into a USING join with the common columns
+        let left_tables = &tables[..table_index];
+        assert!(!left_tables.is_empty());
+        let right_table = &tables[table_index];
+        let right_cols = &right_table.table.columns;
+        let mut distinct_names = None;
+        // TODO: O(n^2) maybe not great for large tables or big multiway joins
+        for right_col in right_cols.iter() {
+            let mut found_match = false;
+            for left_table in left_tables.iter() {
+                for left_col in left_table.table.columns.iter() {
+                    if left_col.name == right_col.name {
+                        if distinct_names.is_none() {
+                            distinct_names =
+                                Some(ast::DistinctNames::new(ast::Name(left_col.name.clone())));
+                        } else {
+                            distinct_names
+                                .as_mut()
+                                .unwrap()
+                                .insert(ast::Name(left_col.name.clone()))
+                                .unwrap();
+                        }
+                        found_match = true;
+                        break;
+                    }
+                }
+                if found_match {
+                    break;
+                }
+            }
+        }
+        if distinct_names.is_none() {
+            crate::bail_parse_error!("No columns found to NATURAL join on");
+        }
+        Some(ast::JoinConstraint::Using(distinct_names.unwrap()))
+    } else {
+        constraint
+    };
+
     if let Some(constraint) = constraint {
         match constraint {
             ast::JoinConstraint::On(expr) => {
@@ -585,7 +648,66 @@ fn parse_join(
                 }
                 predicates = Some(preds);
             }
-            ast::JoinConstraint::Using(_) => todo!("USING joins not supported yet"),
+            ast::JoinConstraint::Using(distinct_names) => {
+                // USING join is replaced with a list of equality predicates
+                let mut using_predicates = vec![];
+                for distinct_name in distinct_names.iter() {
+                    let name_normalized = normalize_ident(distinct_name.0.as_str());
+                    let left_tables = &tables[..table_index];
+                    assert!(!left_tables.is_empty());
+                    let right_table = &tables[table_index];
+                    let mut left_col = None;
+                    for (left_table_idx, left_table) in left_tables.iter().enumerate() {
+                        left_col = left_table
+                            .table
+                            .columns
+                            .iter()
+                            .enumerate()
+                            .find(|(_, col)| col.name == name_normalized)
+                            .map(|(idx, col)| (left_table_idx, idx, col));
+                        if left_col.is_some() {
+                            break;
+                        }
+                    }
+                    if left_col.is_none() {
+                        crate::bail_parse_error!(
+                            "cannot join using column {} - column not present in all tables",
+                            distinct_name.0
+                        );
+                    }
+                    let right_col = right_table
+                        .table
+                        .columns
+                        .iter()
+                        .enumerate()
+                        .find(|(_, col)| col.name == name_normalized);
+                    if right_col.is_none() {
+                        crate::bail_parse_error!(
+                            "cannot join using column {} - column not present in all tables",
+                            distinct_name.0
+                        );
+                    }
+                    let (left_table_idx, left_col_idx, left_col) = left_col.unwrap();
+                    let (right_col_idx, right_col) = right_col.unwrap();
+                    using_predicates.push(ast::Expr::Binary(
+                        Box::new(ast::Expr::Column {
+                            database: None,
+                            table: left_table_idx,
+                            column: left_col_idx,
+                            is_rowid_alias: left_col.primary_key,
+                        }),
+                        ast::Operator::Equals,
+                        Box::new(ast::Expr::Column {
+                            database: None,
+                            table: right_table.table_index,
+                            column: right_col_idx,
+                            is_rowid_alias: right_col.primary_key,
+                        }),
+                    ));
+                }
+                predicates = Some(using_predicates);
+                using = Some(distinct_names);
+            }
         }
     }
 
@@ -597,6 +719,7 @@ fn parse_join(
             iter_dir: None,
         },
         outer,
+        using,
         predicates,
     ))
 }