+++ core/types.rs
@@ -124,7 +124,9 @@ impl PartialOrd<OwnedValue> for OwnedValue {
             (OwnedValue::Null, _) => Some(std::cmp::Ordering::Less),
             (_, OwnedValue::Null) => Some(std::cmp::Ordering::Greater),
             (OwnedValue::Agg(a), OwnedValue::Agg(b)) => a.partial_cmp(b),
-            _ => None,
+            (OwnedValue::Agg(a), other) => a.final_value().partial_cmp(other),
+            (other, OwnedValue::Agg(b)) => other.partial_cmp(b.final_value()),
+            other => todo!("{:?}", other),
         }
     }
 }
@@ -425,6 +427,7 @@ pub enum SeekKey<'a> {
 
 pub trait Cursor {
     fn is_empty(&self) -> bool;
+    fn root_page(&self) -> usize;
     fn rewind(&mut self) -> Result<CursorResult<()>>;
     fn last(&mut self) -> Result<CursorResult<()>>;
     fn next(&mut self) -> Result<CursorResult<()>>;